First attempt at extracting Notes::Tasks
It is probably useful to have programmatic access to all of the tasks
for a file or set of files for potential integration with other
services. Previously, task parsing and printing to the console were
tightly coupled; this extracts the task parsing into the `Notes::Tasks`
module which returns data suitable for formatting and printing.

No functional changes are introduced.