First pass at adding Notes::Tasks.for_files
WIP

We are going to want programmatic access to all of the tasks for a given
project, not just a particular file. This is a start at implementing
that through Notes::Tasks.for_files, which will return a hash of
filename -> [<Notes::Tasks], allowing us to refactor the CLI printing
code as well as display results through other interfaces as desired.