First pass at Backbone-ifying the web interface!
This seems like a pretty decent excuse for learning Backbone - we want a
rich, single-page client-side experience with data models. The intent is
to only load/bootstrap models once, and then perform all
filtering/searching/etc on the client collection.

The dashboard is still crazy slow - this is a mixture of non-optimized
assets as well as searching the file tree on every page request. Some
sort of caching strategy is essential, but not sure how to make it
integrate well with host application. Maybe Redis/Memcached adapter
interface?