Get line ltrimming properly
We want to ltrim lines so that we don't have an excess of indentation
whitespace and everything is nicely aligned. However, we have to take
care to ltrim only to the extent of the smallest amount of leading
whitespace present among both the main line and the context lines.
Consider the following line with 2 additional context lines:

    # TODO: I am an annocation
    Something else here
  end

We can only ltrim those lines by 2, as that is the amount of spaces
preceding the final `end` line.